+++ back-end/reacton/src/main/java/com/softeer/reacton/global/jwt/filter/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.softeer.reacton.global.jwt;
+package com.softeer.reacton.global.jwt.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.softeer.reacton.global.exception.BaseException;
