+++ back-end/reacton/src/main/java/com/softeer/reacton/global/jwt/resolver/ProfessorAuthResolver.java
@@ -1,12 +1,14 @@
-package com.softeer.reacton.global.jwt;
+package com.softeer.reacton.global.jwt.resolver;
 
 import com.softeer.reacton.global.exception.BaseException;
 import com.softeer.reacton.global.exception.code.JwtErrorCode;
-import com.softeer.reacton.global.jwt.dto.StudentAuthInfo;
+import com.softeer.reacton.global.jwt.util.JwtTokenUtil;
+import com.softeer.reacton.global.jwt.dto.ProfessorAuthInfo;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -14,15 +16,16 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
-public class StudentAuthResolver implements HandlerMethodArgumentResolver {
+public class ProfessorAuthResolver implements HandlerMethodArgumentResolver {
 
     private final JwtTokenUtil jwtTokenUtil;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterType().equals(StudentAuthInfo.class);
+        return parameter.getParameterType().equals(ProfessorAuthInfo.class);
     }
 
     @Override
@@ -36,20 +39,19 @@ public class StudentAuthResolver implements HandlerMethodArgumentResolver {
         }
 
         Claims claims = jwtTokenUtil.getClaims(token);
-        String studentId = claims.get("studentId", String.class);
-        Long courseId = claims.get("courseId", Long.class);
-
-        if (studentId == null || courseId == null) {
+        Integer professorIdInt = claims.get("professorId", Integer.class);
+        if (professorIdInt == null) {
             throw new BaseException(JwtErrorCode.ACCESS_TOKEN_ERROR);
         }
 
-        return new StudentAuthInfo(studentId, courseId);
+        Long professorId = professorIdInt.longValue();
+        return new ProfessorAuthInfo(professorId);
     }
 
     private String extractTokenFromRequest(HttpServletRequest request) {
         if (request.getCookies() != null) {
             for (Cookie cookie : request.getCookies()) {
-                if ("student_access_token".equals(cookie.getName())) {
+                if ("access_token".equals(cookie.getName())) {
                     return cookie.getValue();
                 }
             }
