+++ back-end/reacton/src/main/java/com/softeer/reacton/global/jwt/resolver/StudentAuthResolver.java
@@ -1,13 +1,13 @@
-package com.softeer.reacton.global.jwt;
+package com.softeer.reacton.global.jwt.resolver;
 
 import com.softeer.reacton.global.exception.BaseException;
 import com.softeer.reacton.global.exception.code.JwtErrorCode;
-import com.softeer.reacton.global.jwt.dto.ProfessorAuthInfo;
+import com.softeer.reacton.global.jwt.util.JwtTokenUtil;
+import com.softeer.reacton.global.jwt.dto.StudentAuthInfo;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -15,16 +15,15 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-@Slf4j
 @Component
 @RequiredArgsConstructor
-public class ProfessorAuthResolver implements HandlerMethodArgumentResolver {
+public class StudentAuthResolver implements HandlerMethodArgumentResolver {
 
     private final JwtTokenUtil jwtTokenUtil;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterType().equals(ProfessorAuthInfo.class);
+        return parameter.getParameterType().equals(StudentAuthInfo.class);
     }
 
     @Override
@@ -38,19 +37,20 @@ public class ProfessorAuthResolver implements HandlerMethodArgumentResolver {
         }
 
         Claims claims = jwtTokenUtil.getClaims(token);
-        Integer professorIdInt = claims.get("professorId", Integer.class);
-        if (professorIdInt == null) {
+        String studentId = claims.get("studentId", String.class);
+        Long courseId = claims.get("courseId", Long.class);
+
+        if (studentId == null || courseId == null) {
             throw new BaseException(JwtErrorCode.ACCESS_TOKEN_ERROR);
         }
 
-        Long professorId = professorIdInt.longValue();
-        return new ProfessorAuthInfo(professorId);
+        return new StudentAuthInfo(studentId, courseId);
     }
 
     private String extractTokenFromRequest(HttpServletRequest request) {
         if (request.getCookies() != null) {
             for (Cookie cookie : request.getCookies()) {
-                if ("access_token".equals(cookie.getName())) {
+                if ("student_access_token".equals(cookie.getName())) {
                     return cookie.getValue();
                 }
             }
